Move ResponseGenerator to listeners package.
