+++ generator/src/main/java/com/algorand/sdkutils/listeners/JavaGenerator.java
@@ -1,0 +1,4 @@
+package com.algorand.sdkutils.listeners;
+
+public class JavaGenerator {
+}
